+++ authentication-service/src/main/java/nl/markpost/authentication/dto/UpdateUserNameRequest.java
@@ -1,0 +1,6 @@
+package nl.markpost.authentication.dto;
+
+public record UpdateUserNameRequest(String username) {
+
+}
+